feat(find): SUI-1773 - Map client id to organisation id in JWT middleware
